refactor: 검색 페이지 레이어 분리 및 regex 호이스팅 (#69)
