+++ src/features/epigram-search/index.ts
@@ -1,4 +1,6 @@
 export { useRecentSearches } from "./model/useRecentSearches";
 export { useSearch } from "./model/useSearch";
 export { SearchBar } from "./ui/SearchBar";
+export { RecentSearchList } from "./ui/RecentSearchList";
 export { SearchResultItem } from "./ui/SearchResultItem";
+export { SearchResults } from "./ui/SearchResults";
